# codegen/openapi.yaml
openapi: 3.1.0
info:
  title: RPA Control-Plane API
  version: "1.0"
  description: >
    제어평면 REST API (v1). 운영 콘솔/외부 클라이언트가 호출하는 제어평면 엔드포인트의
    경로·메서드·요청/응답 요지·주요 에러코드를 고정한다. 본 문서는 api-surface.md(Control-Plane
    API Surface v1)를 1:1 변환한 D1 codegen 산출물이다.


    요지 수준 규약: components.schemas는 요지(shape) 수준이며, 상세 타입(IR/Verify/Event payload·필드
    required·examples)은 계약 ts/schema 및 codegen/types를 참조한다. 어휘 체인 abort → cancelled →
    run.cancelled, "조용한 false/unknown 금지", error-catalog 메타(httpStatus/exceptionClass)를 그대로 따른다.


    - 인증: 모든 엔드포인트 인증 필수. tenant_id는 요청 본문/쿼리가 아니라 인증 주체(토큰)에서 도출한다(api-surface §0.1).

    - 에러 본문: 모든 4xx/5xx는 ApiError(ts/error-catalog.ts). HTTP 상태코드는 ERROR_CATALOG[code].httpStatus를 그대로 사용한다(api-surface §0.2).

    - DEAD_LETTER(httpStatus 200)는 API 오류 응답이 아니다 — 상태 통지 전용 코드이므로 ApiError로 반환하지 않는다(api-surface §0.2/§4).
servers:
  - url: /v1
    description: Control-plane v1 base path.
...
security:
  - bearerAuth: []
paths:
...
  /document-jobs:
    get:
      tags: [DocumentIDP]
      operationId: listDocumentJobs
      summary: List document automation jobs
      description: List tenant-scoped document extraction jobs. Source document bytes are never returned by this endpoint.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/DocumentJobStatus'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Document job page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentJobPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [DocumentIDP]
      operationId: createDocumentJob
      summary: Create a document extraction job
      description: Create a deterministic text/CSV/JSON document extraction job from a redacted source artifact.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentJobCreateRequest'
      responses:
        '201':
          description: Created document job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentJob'
        default:
          $ref: '#/components/responses/Error'
  /document-jobs/{job_id}:
    get:
      tags: [DocumentIDP]
      operationId: getDocumentJob
      summary: Get a document extraction job
      description: Return a tenant-scoped document job without source artifact bytes.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Document job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentJob'
        default:
          $ref: '#/components/responses/Error'
  /document-jobs/{job_id}/extract:
    post:
      tags: [DocumentIDP]
      operationId: extractDocumentJob
      summary: Run deterministic document extraction
      description: Read the redacted source artifact through the artifact audit boundary and store extraction fields.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentJobCommandRequest'
      responses:
        '200':
          description: Document extraction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExtraction'
        default:
          $ref: '#/components/responses/Error'
  /document-jobs/{job_id}/extraction:
    get:
      tags: [DocumentIDP]
      operationId: getDocumentExtraction
      summary: Get stored document extraction result
      description: Return stored extraction fields and validation state for a document job.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Document extraction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExtraction'
        default:
          $ref: '#/components/responses/Error'
  /document-jobs/{job_id}/validation-task:
    post:
      tags: [DocumentIDP]
      operationId: createDocumentValidationTask
      summary: Create a human validation task for missing or low-confidence fields
      description: Create or return the reviewer task that captures corrected document field values.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentJobCommandRequest'
      responses:
        '201':
          description: Human task reference for validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentValidationTask'
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT bearer. tenant_id 및 roles(Role[]) 클레임에서 도출(auth-rbac §3). 클라이언트 body/query의
        tenant_id는 불신. 인증 미성립(토큰 누락/서명 무효) 시 UNAUTHENTICATED(401), 인증됐으나 tenant
        누락/모호 또는 역할 권한 부족 시 AUTHZ_FORBIDDEN(403). 두 경우 모두 본문은 ApiError.
  # --- Reusable parameters ---
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: 페이지 크기. 상한 기본값은 운영 정책(Phase 3) — 본 계약은 파라미터 형태만 고정(api-surface §0.5).
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 불투명 커서. 직전 응답 next_cursor 값을 그대로 전달(api-surface §0.5).
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        부작용 명령형 POST의 중복 제출 보호(api-surface §0.4). 서버는 (tenant_id, endpoint, key)로
        최초 결과를 보관, 동일 키 재제출 시 부작용 재실행 없이 최초 응답 반환. sink_idempotency_key(데이터평면)와는 다른 계층.
      schema:
        type: string

  # --- Reusable headers ---
...
  responses:
    Error:
      description: >
        오류 응답. 본문은 ApiError(ts/error-catalog.ts). HTTP 상태는 ERROR_CATALOG[code].httpStatus를
        그대로 사용한다. DEAD_LETTER는 오류 응답이 아니므로 본문 code로 반환되지 않는다(api-surface §0.2).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # --- Schemas (요지 수준; 상세 타입은 계약 ts/schema 및 codegen/types 참조) ---
  schemas:
...
    ApiError:
      type: object
      description: 공통 에러 응답(ts/error-catalog.ts ApiError).
      required: [code, message, correlation_id]
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: userMessage 또는 상세(민감정보 없음).
        details:
          description: 코드별 상세 페이로드(예 IR_SCHEMA_INVALID의 reason). 구조는 D1 위임(api-surface §9).
        correlation_id:
          type: string
          format: uuid
          description: event-envelope·trace correlation_id와 동일 값(api-surface §0.2).
    ErrorCode:
      type: string
      description: >
        ts/error-catalog.ts ErrorCode(48개). DEAD_LETTER는 enum에 포함되나 상태 통지 전용으로
        ApiError 응답 code로는 반환되지 않는다(api-surface §0.2).
      enum:
        - RUN_NOT_FOUND
        - RESOURCE_NOT_FOUND
        - RUN_ALREADY_TERMINAL
        - RUN_ABORTED
        - SCENARIO_VERSION_CONFLICT
        - POLICY_VERSION_CONFLICT
        - IR_SCHEMA_INVALID
        - IR_EXPRESSION_COMPILE_ERROR
        - IR_EXPRESSION_RUNTIME
        - IR_NO_BRANCH_MATCHED
        - SITE_PROFILE_BLOCKED
        - SITE_CIRCUIT_OPEN
        - SESSION_LOCKED
        - SESSION_GENERATION_CONFLICT
        - SESSION_REGISTRATION_REQUIRED
        - CHALLENGE_UNRESOLVED
        - RATE_BUDGET_EXCEEDED
        - BROWSER_LEASE_EXPIRED
        - BROWSER_CRASH
        - CDP_DISCONNECTED
        - NAVIGATION_TIMEOUT
        - LLM_BUDGET_EXCEEDED
        - LLM_CAPABILITY_MISMATCH
        - LLM_STREAM_TIMEOUT
        - LLM_STREAM_IDLE_TIMEOUT
        - LLM_MALFORMED_OUTPUT
        - LLM_CONTENT_FILTERED
        - LLM_RATE_LIMITED
        - LLM_BACKEND_UNAVAILABLE
        - LLM_CONNECTION_FAILED
        - EXTRACT_SCHEMA_INVALID
        - VERIFY_FAILED
        - EMPTY_RESULT_NO_WITNESS
        - SECRET_ACCESS_DENIED
        - DOMAIN_POLICY_VIOLATION
        - PROMPT_INJECTION_DETECTED
        - ARTIFACT_NOT_REDACTED
        - SHELL_COMMAND_NOT_ALLOWED
        - UNAUTHENTICATED
        - AUTHZ_FORBIDDEN
        - CONNECTOR_PERMISSION_DENIED
        - CONNECTOR_INCOMPATIBLE
        - CONNECTOR_HOOK_FAILED
        - SINK_DELIVERY_FAILED
        - RAW_PERSIST_FAILED
        - CONTROL_PLANE_INTERNAL_ERROR
        - HUMAN_TASK_EXPIRED
        - APPROVAL_ALREADY_DECIDED
        - WORKITEM_CHECKOUT_CONFLICT
        - DEAD_LETTER

    # ===== State enums (ts/state-machine-types.ts — 정확히 일치) =====
...
    DocumentJobStatus:
      type: string
      enum:
        - created
        - extracted
        - validation_required
        - validated
        - failed
    DocumentExtractionStatus:
      type: string
      enum:
        - completed
        - validation_required
        - failed
    DocumentFieldType:
      type: string
      enum:
        - text
        - number
        - date
        - boolean
...
    HumanTaskState:
      type: string
      enum:
        - open
        - assigned
        - in_progress
        - resolved
        - expired
        - cancelled
        - escalated
...
    DocumentJobPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/DocumentJob' }
        next_cursor:
          type: [string, "null"]
...
    DocumentFieldSchema:
      type: object
      additionalProperties: false
      required: [key]
      properties:
        key: { type: string, minLength: 1, maxLength: 80 }
        label: { type: string, minLength: 1 }
        type:
          $ref: '#/components/schemas/DocumentFieldType'
        required: { type: boolean, default: false }
        aliases:
          type: array
          items: { type: string, minLength: 1 }
        patterns:
          type: array
          description: JavaScript regular expressions with optional capture group 1 as extracted value.
          items: { type: string, minLength: 1 }
        min_confidence:
          type: number
          minimum: 0
          maximum: 1
    DocumentJobCreateRequest:
      type: object
      additionalProperties: false
      required: [source_artifact_id, document_type, field_schema]
      properties:
        source_artifact_id: { type: string, format: uuid }
        document_type: { type: string, minLength: 1, maxLength: 120 }
        field_schema:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/DocumentFieldSchema'
    DocumentJobCommandRequest:
      type: object
      additionalProperties: false
    DocumentJob:
      type: object
      additionalProperties: false
      required:
        - document_job_id
        - source_artifact_id
        - source_run_id
        - document_type
        - field_schema
        - status
        - created_by
        - created_at
        - updated_at
      properties:
        document_job_id: { type: string, format: uuid }
        source_artifact_id: { type: string, format: uuid }
        source_run_id: { type: string, format: uuid }
        document_type: { type: string }
        field_schema:
          type: array
          items:
            $ref: '#/components/schemas/DocumentFieldSchema'
        status:
          $ref: '#/components/schemas/DocumentJobStatus'
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DocumentExtractionField:
      type: object
      additionalProperties: false
      required: [key, label, value, confidence, status, source]
      properties:
        key: { type: string }
        label: { type: string }
        value: { type: [string, "null"] }
        confidence:
          type: number
          minimum: 0
          maximum: 1
        status:
          type: string
          enum: [extracted, missing, low_confidence]
        source:
          type: string
          enum: [json, csv, pattern, label, missing]
    DocumentExtraction:
      type: object
      additionalProperties: false
      required:
        - document_extraction_id
        - document_job_id
        - engine
        - status
        - fields
        - missing_fields
        - validation_human_task_id
        - created_at
        - updated_at
      properties:
        document_extraction_id: { type: string, format: uuid }
        document_job_id: { type: string, format: uuid }
        engine:
          type: string
          enum: [built_in_deterministic_text_v1]
        status:
          $ref: '#/components/schemas/DocumentExtractionStatus'
        fields:
          type: array
          items:
            $ref: '#/components/schemas/DocumentExtractionField'
        missing_fields:
          type: array
          items: { type: string }
        validation_human_task_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DocumentValidationTask:
      type: object
      additionalProperties: false
      required: [human_task_id, state, result_schema, artifact_refs]
      properties:
        human_task_id: { type: string, format: uuid }
        state:
          $ref: '#/components/schemas/HumanTaskState'
        result_schema:
          type: object
          description: business_form_v1 schema for reviewer correction input.
          additionalProperties: true
        artifact_refs:
          type: array
          items: { type: string }
...